Use solver_options to specify solver-dependent options
Options that are shared by multiple solvers are handled as keywords
arguments, but solver-dependent options are defined in the
solver_options dictionary:

  Solver(name='picosat', generate_models=True,
         solver_options={'propagation_limit': 100})

The goal is to clearly separate these two type of configurations.
The use of dictionary for solver_options, makes it simple to store
multiple configurations:

  Solver(name=sname, solver_options=myoptions[sname])

All solvers have been ported to this new strategy, by using a custom
SolverOptions object that is responsible for the definition, checking
and setting of the options. This object is used (for example) to
correctly setup the mathsat context.

Option "random_seed" is now supported by all the solvers (that do
support it).

Deprecation:

Backwards Incompatible Changes:

* MathSAT5Solver option 'debugFile' has been removed deprecated. Use the
  solver option: "debug_api_call_trace_filename".

* BddSolver used to have the options as keyword
  arguments (static_ordering, dynamic_reordering etc). This is not
  supported anymore.